refactor(cockpit): move proof evidence tests out of coordinator
